Move filter helper to service
