+++ controllers/api_v1/keyword.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"github.com/gutakk/go-google-scraper/helpers/api_helper"
-	"github.com/gutakk/go-google-scraper/helpers/filter_helper"
 	helpers "github.com/gutakk/go-google-scraper/helpers/user"
 	"github.com/gutakk/go-google-scraper/serializers"
 	"github.com/gutakk/go-google-scraper/services/keyword_service"
@@ -26,8 +25,8 @@ func (kapi *KeywordAPIController) ApplyRoutes(engine *gin.RouterGroup) {
 
 func (kapi *KeywordAPIController) fetchKeywords(c *gin.Context) {
 	currentUserID := helpers.GetCurrentUserID(c)
-	keywordService := keyword_service.KeywordService{CurrentUserID: currentUserID}
-	conditions := filter_helper.FilterValidConditions(c.Request.URL.Query())
+	keywordService := keyword_service.KeywordService{CurrentUserID: currentUserID, QueryString: c.Request.URL.Query()}
+	conditions := keywordService.FilterValidConditions()
 	keywords, err := keywordService.GetKeywords(conditions)
 
 	if err != nil {
