+++ services/keyword_service/base.go
@@ -1,0 +1,6 @@
+package keyword_service
+
+type KeywordService struct {
+	CurrentUserID uint
+	QueryString   map[string][]string
+}
